+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2024-12-08
+
+### Added
+
+- TSID support.
+- Sqids support.
+- Support custom alphabet for Sqids and Nono ID.
+
+### Changed
+
+- UUID node (for versions 1 and 6) are now formatted like a MAC-address.
+
 ## [0.1.1] - 2024-12-05
 
 ### Changed